requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
langchain>=0.0.350
openai>=1.3.0
anthropic>=0.5.0
gitpython>=3.1.40
pydantic>=2.5.0
python-multipart>=0.0.6
requests>=2.31.0
PyGithub>=2.1.1
python-gitlab>=3.15.0 